Unicorn/requirements.txt
numpy>=1.21.0
pynput>=1.7.6
keyboard>=0.13.5
unicorn-py>=1.0.0
matplotlib>=3.5.0
pillow>=9.0.0 